feat: add new flag package for posix flag parsing
This revision introduces a new package `flag` similar to the standard
library package of the same name which parses POSIX/GNU style options
arguments. The package is a 1-for-1 implementation of the standard
package but with an updated parser.

At the moment, cmder isn't configure to use it. This will be introduced
in a future revision.